execution/requirements.txt
pandas
streamlit
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
streamlit-aggrid
google-cloud-storage
playwright
keyring
